@@ -11,9 +11,9 @@ use Symfony\Component\Routing\Annotation\Route;
 
 class LegacyController extends AbstractController
 {
-    /**
-     * @Route("/.htaccess", name="htaccess", format="htaccess", options={"stenope"={"ignore"=true}})
-     */
+    #[Route('/.htaccess', name: 'htaccess', options: [
+        'stenope' => ['ignore' => true],
+    ], format: 'htaccess')]
     public function htaccess(HtaccessGenerator $generator): Response
     {
         return new Response($generator->generateHtAccess(HtaccessGenerator::TARGET_SITE));
